+++ lib/spdylay_map.c
@@ -96,7 +96,7 @@ static int32_t hash(int32_t h, int32_t mod)
 static int insert(spdylay_map_entry **table, size_t tablelen,
                   spdylay_map_entry *entry)
 {
-  int32_t h = hash(entry->key, tablelen);
+  int32_t h = hash(entry->key, (int32_t)tablelen);
   if(table[h] == NULL) {
     table[h] = entry;
   } else {
@@ -161,7 +161,7 @@ spdylay_map_entry* spdylay_map_find(spdylay_map *map, key_type key)
 {
   int32_t h;
   spdylay_map_entry *entry;
-  h = hash(key, map->tablelen);
+  h = hash(key, (int32_t)map->tablelen);
   for(entry = map->table[h]; entry; entry = entry->next) {
     if(entry->key == key) {
       return entry;
@@ -174,7 +174,7 @@ int spdylay_map_remove(spdylay_map *map, key_type key)
 {
   int32_t h;
   spdylay_map_entry *entry, *prev;
-  h = hash(key, map->tablelen);
+  h = hash(key, (int32_t)map->tablelen);
   prev = NULL;
   for(entry = map->table[h]; entry; entry = entry->next) {
     if(entry->key == key) {
