+++ tests/spdylay_frame_test.c
@@ -821,11 +821,11 @@ static size_t spdylay_pack_nv(uint8_t *buf, size_t buflen, const char **nv,
   uint8_t *buf_ptr;
   buf_ptr = buf;
   for(n = 0; nv[n]; ++n);
-  spdylay_frame_put_nv_len(buf_ptr, n/2, len_size);
+  spdylay_frame_put_nv_len(buf_ptr, (uint32_t)(n/2), len_size);
   buf_ptr += len_size;
   for(i = 0; i < n; ++i) {
     size_t len = strlen(nv[i]);
-    spdylay_frame_put_nv_len(buf_ptr, len, len_size);
+    spdylay_frame_put_nv_len(buf_ptr, (uint32_t)len, len_size);
     buf_ptr += len_size;
     memcpy(buf_ptr, nv[i], len);
     buf_ptr += len;
